consumer/requirements.txt
flask
kafka-python
sqlalchemy
psycopg2-binary
python-dotenv
mysql-connector-python
pymysql
